moved locksymbol to own class, added layouting in all presetlayouters and refactored
