+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/world/overlay/belt/SoundBeltButton.java
@@ -10,8 +10,6 @@ import com.nonlinearlabs.NonMaps.client.world.overlay.OverlayLayout;
 
 public class SoundBeltButton extends OverlayLayout implements IActivatable {
 
-	private static double lockSymbolSizeInMM = 4;
-
 	private class SoundIcon extends BeltButton {
 
 		private SoundIcon(OverlayLayout parent, Belt belt) {
@@ -34,36 +32,17 @@ public class SoundBeltButton extends OverlayLayout implements IActivatable {
 		}
 	}
 
-	private class LockSymbol extends Label {
-
-		public LockSymbol(OverlayLayout parent) {
-			super(parent);
-			setFontHeightInMM(lockSymbolSizeInMM);
-		}
-
-		@Override
-		public String getDrawText(Context2d ctx) {
-			if (NonMaps.get().getNonLinearWorld().getParameterEditor().isAnyParameterLocked())
-				return "\ue20a";
-			return "";
-		}
-	}
-
 	private SoundIcon icon;
-	private LockSymbol lock;
 
 	public SoundBeltButton(OverlayLayout parent, Belt belt) {
 		super(parent);
 		icon = addChild(new SoundIcon(this, belt));
-		lock = addChild(new LockSymbol(this));
 	}
 
 	@Override
 	public void doLayout(double x, double y, double w, double h) {
 		super.doLayout(x, y, w, h);
 		icon.doLayout(0, 0, w, h);
-		double lockSymbolSizeInPixel = Millimeter.toPixels(lockSymbolSizeInMM + 1);
-		lock.doLayout(w - lockSymbolSizeInPixel, 0, lockSymbolSizeInPixel, lockSymbolSizeInPixel);
 	}
 
 	@Override
