+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/world/overlay/belt/presets/PresetBeltLayouter.java
@@ -14,6 +14,11 @@ public class PresetBeltLayouter {
 		return false;
 	}
 
+	protected void layoutLockSymbol(double w) {
+		double lockSymbolSizeInPixel = Millimeter.toPixels(layout.lock.getLockSizeInMM() + 1);
+		layout.lock.doLayout(w - lockSymbolSizeInPixel, 0, lockSymbolSizeInPixel, lockSymbolSizeInPixel);
+	}
+
 	protected void hide(OverlayControl ctrl) {
 		ctrl.doLayout(0, 0, 0, 0);
 	}
