+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/world/overlay/belt/presets/PresetBeltLayouterXS.java
@@ -30,6 +30,7 @@ public class PresetBeltLayouterXS extends PresetBeltLayouter {
 
 		layout.bank.doLayout(left, 0, bankWidth, h);
 
+		super.layoutLockSymbol(w);
 		return true;
 	}
 }
