+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/world/overlay/belt/presets/PresetBeltLayouterXXL.java
@@ -56,6 +56,7 @@ public class PresetBeltLayouterXXL extends PresetBeltLayouter {
 		right += layout.load.getRelativePosition().getWidth() + margin;
 		layout.currentPreset.getRelativePosition().moveTo(right, layout.currentPreset.getRelativePosition().getTop());
 
+		super.layoutLockSymbol(w);
 		return true;
 	}
 }
