+++ NonMaps/src/com/nonlinearlabs/NonMaps/client/world/overlay/belt/sound/BeltSoundLayout.java
@@ -2,15 +2,14 @@ package com.nonlinearlabs.NonMaps.client.world.overlay.belt.sound;
 
 import com.google.gwt.xml.client.Node;
 import com.nonlinearlabs.NonMaps.client.Millimeter;
-import com.nonlinearlabs.NonMaps.client.NonMaps;
-import com.nonlinearlabs.NonMaps.client.world.overlay.OverlayControl;
-import com.nonlinearlabs.NonMaps.client.world.overlay.OverlayLayout;
-import com.nonlinearlabs.NonMaps.client.world.overlay.belt.Belt;
+import com.nonlinearlabs.NonMaps.client.world.overlay.*;
+import com.nonlinearlabs.NonMaps.client.world.overlay.belt.*;
 
 public class BeltSoundLayout extends OverlayLayout {
 
 	private TransitionTimeArea transitionTime;
 	private RandomizeSoundArea randomize;
+	private LockSymbol lock;
 
 	public BeltSoundLayout(Belt belt) {
 		super(belt);
@@ -19,6 +18,7 @@ public class BeltSoundLayout extends OverlayLayout {
 		addChild(randomize = new RandomizeSoundArea(this));
 		addChild(transitionTime = new TransitionTimeArea(this));
 		addChild(new ConvertSoundArea(this));
+		lock = addChild(new LockSymbol(this));
 	}
 
 	@Override
@@ -42,13 +42,15 @@ public class BeltSoundLayout extends OverlayLayout {
 			xPos += rest / 2;
 		}
 
-		double height = h;
-		double yPos = (h - height) / 2;
+		double yPos = 0;
 
 		for (OverlayControl c : getChildren()) {
-			c.doLayout(xPos, yPos, widthPerArea, height);
+			c.doLayout(xPos, yPos, widthPerArea, h);
 			xPos += widthPerArea + margin;
 		}
+
+		double lockSymbolSizeInPixel = Millimeter.toPixels(lock.getLockSizeInMM() + 1);
+		lock.doLayout(w - lockSymbolSizeInPixel, 0, lockSymbolSizeInPixel, lockSymbolSizeInPixel);
 	}
 
 	public void update(Node settings) {
